refactor: move proposal list fetching to react-query
